backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0
requests>=2.31.0
python-jose[cryptography]>=3.3.0
stripe>=7.0.0
alembic>=1.12.0
slowapi>=0.1.9
